requirements.txt
accelerate
torchvision
transformers>=4.25.1
ftfy
tensorboard
Jinja2
gradio